backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Admin==1.6.0
Flask-Login==0.5.0
Flask-WTF==1.0.0
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
python-dotenv==0.19.1
gunicorn==20.1.0
